backend/requirements.txt
fastapi==0.124.4
uvicorn==0.38.0
python-multipart==0.0.20
pydantic==2.12.5
numpy==2.3.5
librosa==0.11.0
scikit-learn==1.8.0
soundfile==0.13.1
supabase==2.28.3
python-dotenv==1.2.2
